chore: public promotion cleanup and open-source hardening (#99)
* chore: public promotion cleanup


* fix: eliminate rustls-webpki vulnerabilities via AWS SDK feature pruning

aws-sdk-s3's default  feature enabled ,
which pulled in the legacy rustls 0.21 stack (hyper-rustls 0.24 → tokio-rustls
0.24 → rustls 0.21 → rustls-webpki 0.101.7). This version has three active
RUSTSEC advisories (0098, 0099, 0104).

By disabling default features and explicitly enabling only the modern HTTPS
path ( →  → rustls 0.23), Cargo drops the
entire legacy rustls 0.21 dependency tree from the lockfile.

Also bumps push-proxy lockfile to pull rustls-webpki 0.103.13 (fixes
RUSTSEC-2026-0104 on the rustls 0.23 branch).

Refs: #99

* fix(playbooks): resolve pending tasks and run detail failures

- Remove #[serde(flatten)] from RunWithTasks to match frontend nested
  expectation ({ run, tasks, progress })
- Fix RunWithTasks TypeScript interface (remove erroneous top-level id)
- Fix update_run_task SQL: clear completed_at/completed_by when status
  changes from done to non-done using CASE expressions instead of COALESCE
- Fix optimistic update revert in PlaybookRun.vue: restore progress counters
  on API failure
- Fix playbookRepository/playbookService calling wrong endpoint
  (updateTask -> updateRunTask) for run task updates


* ci: add legacy status check aggregators for branch protection compatibility

The branch protection rules on main require status checks named:
- Build
- cargo check + test
- Test

These names came from the old ci.yml + backend-ci.yml workflow structure.
When workflows were consolidated, the job names changed but branch
protection was not updated.

Add three aggregator jobs that always run (if: always()) and report
success/failure under the legacy names, mapping to the new jobs:
- Test → backend-check + frontend-check + push-proxy-check
- Build → backend-check + build-release
- cargo check + test → backend-check

---------

+++ backend/src/api/admin.rs
@@ -1558,9 +1558,13 @@ fn check_disk(path: &str) -> DiskHealth {
             }
         };
         if libc::statvfs(c_path.as_ptr(), &mut stat) == 0 {
-            let block_size = stat.f_frsize;
-            let total = stat.f_blocks * block_size;
-            let available = stat.f_bavail * block_size;
+            // Cast needed for cross-platform compatibility (types vary by OS/arch)
+            #[allow(clippy::unnecessary_cast)]
+            let block_size = stat.f_frsize as u64;
+            #[allow(clippy::unnecessary_cast)]
+            let total = stat.f_blocks as u64 * block_size;
+            #[allow(clippy::unnecessary_cast)]
+            let available = stat.f_bavail as u64 * block_size;
             let used = total.saturating_sub(available);
             let used_percent = if total > 0 {
                 ((used as f64 / total as f64) * 100.0) as u64
@@ -1570,7 +1574,7 @@ fn check_disk(path: &str) -> DiskHealth {
             DiskHealth {
                 connected: true,
                 used_percent,
-                available_mb: available / (1024 * 1024),
+                available_mb: available / (1024u64 * 1024u64),
             }
         } else {
             DiskHealth {
